Start to split the too complex forward function
Extract a function that collects to toots to be sent.

It returns a list of toot descriptions, what ever the
`only_mark_as_seen` parameter value was passed to the forwarding
function. The use of this variable is left to the forward function,
which improves the readability.

More simple refactorings are coming up to simplify the forward function
and make the whole code more easily unit-testable (and to add tests at
the end).