Added suport for automatically adding content warnings
This change adds support for configuring rules in a JSON document for
content warnings that should be applied when posting a toot to Mastodon.

The rules can either be a simple regex match against the text of a
tweet, or can include a group to allow using part of the text of the
tweet to be the text of the content warning.

The patterns are compared after applying the retweet.tmpl template which
allows things like automatically adding a content warning that specifies
the name of the twitter user being RTed.

The README.md has also been updated witn usage instructions and examples